+++ rock/src/rock.mli
@@ -11,4 +11,3 @@ module Service = Service
 module Filter = Filter
 module Handler = Handler
 module Middleware = Middleware
-module Server_connection = Server_connection